FEATURE: Using AspNet project organization pattern
