+++ src/Arriba/Arriba.Server/appsettings.json
@@ -8,14 +8,14 @@
   },
   "AllowedHosts": "*",
   "ArribaServer": {
-    "EnabledAuthentication": true,
-    "FrontendBaseUrl": "https://arriba-search.azurewebsites.net/",
+    "EnabledAuthentication": false,
+    "FrontendBaseUrl": "https://localhost:8080/",
     "OAuthConfig": {
-      "RedirectUrl": "https://arriba-search.azurewebsites.net/data/api/oauth/auth-code",
-      "TenantId": "72f988bf-86f1-41af-91ab-2d7cd011db47",
-      "AudienceId": "ae4ffb67-b66b-4b8f-907b-9e64e5249199",
+      "RedirectUrl": "https://localhost:8080/data/api/oauth/auth-code",
+      "TenantId": "<AZURE AD TENANT ID>",
+      "AudienceId": "<AZURE AD APP CLIENT ID>",
       "Prompt": "login",
-      "Scopes": [ "openid", "ae4ffb67-b66b-4b8f-907b-9e64e5249199/.default" ]
+      "Scopes": [ "openid", "<AZURE AD APP CLIENT ID>/.default" ]
     }
   }
 }
